requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
click==8.1.6
colorama==0.4.6
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
flake8==6.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.2.0
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
